add options; refactor test code

+++ src/main/java/com/ma/cm/service/ColumnService.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.ma.cm.entity.Column;
+import com.ma.cm.entity.ColumnContent;
+import com.ma.cm.exception.ForbiddenException;
 
 @Service
 public class ColumnService extends AService {
@@ -43,6 +45,11 @@ public class ColumnService extends AService {
 	@Transactional
 	public void delete(long productId, long columnId) {
 		checkColumnExist(productId, columnId);
+		List<ColumnContent> columnContents = columnContentMapper.getByProductByContentByType1(productId, columnId);
+		if (!columnContents.isEmpty()) {
+			String message = String.format("should not delete the column %d:%d because it is used by column %d:%d", productId, columnId, productId, columnContents.get(0).getColumnId());
+			throw new ForbiddenException(message);
+		}
 		columnContentMapper.deleteByProductByColumn(productId, columnId);
 		columnMapper.delete(productId, columnId);
 	}
